+++ app/Http/Controllers/LocationController.php
@@ -1,12 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Controllers;
 
 use App\Http\Requests\IndexLocationRequest;
 use App\Http\Requests\StoreLocationRequest;
 use App\Http\Requests\UpdateLocationRequest;
+use App\Http\Resources\LocationCollection;
 use App\Http\Resources\LocationResource;
+use App\Models\Location;
 use App\Services\LocationService;
+use Illuminate\Http\JsonResponse;
 
 class LocationController extends Controller
 {
@@ -16,12 +21,15 @@ class LocationController extends Controller
      * @param \App\Http\Requests\IndexLocationRequest $request
      * @param \App\Services\LocationService $service
      * @return \App\Http\Resources\LocationCollection
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function index(IndexLocationRequest $request, LocationService $service)
+    public function index(IndexLocationRequest $request, LocationService $service): LocationCollection
     {
-        $locations = $service->getAllLocations();
+        $this->authorize('viewAny', Location::class);
 
-        return LocationResource::collection($locations);
+        $locations = $service->getLocationsByUser(auth()->id());
+
+        return new LocationCollection($locations);
     }
 
     /**
@@ -30,9 +38,12 @@ class LocationController extends Controller
      * @param \App\Http\Requests\StoreLocationRequest $request
      * @param \App\Services\LocationService $service
      * @return \App\Http\Resources\LocationResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function store(StoreLocationRequest $request, LocationService $service)
+    public function store(StoreLocationRequest $request, LocationService $service): LocationResource
     {
+        $this->authorize('create', Location::class);
+
         $location = $service->createLocation($request->validated());
 
         return new LocationResource($location);
@@ -44,9 +55,12 @@ class LocationController extends Controller
      * @param string $locationId
      * @param \App\Services\LocationService $service
      * @return \App\Http\Resources\LocationResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function show(string $locationId, LocationService $service)
+    public function show(string $locationId, LocationService $service): LocationResource
     {
+        $this->authorize('view', [Location::class, $locationId]);
+
         $location = $service->getLocationById($locationId);
 
         return new LocationResource($location);
@@ -59,9 +73,12 @@ class LocationController extends Controller
      * @param string $locationId
      * @param \App\Services\LocationService $service
      * @return \App\Http\Resources\LocationResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function update(UpdateLocationRequest $request, string $locationId, LocationService $service)
+    public function update(UpdateLocationRequest $request, string $locationId, LocationService $service): LocationResource
     {
+        $this->authorize('update', [Location::class, $locationId]);
+
         $location = $service->updateLocation($request->validated(), $locationId);
 
         return new LocationResource($location);
@@ -73,9 +90,12 @@ class LocationController extends Controller
      * @param string $locationId
      * @param \App\Services\LocationService $service
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function destroy(string $locationId, LocationService $service)
+    public function destroy(string $locationId, LocationService $service): JsonResponse
     {
+        $this->authorize('delete', [Location::class, $locationId]);
+
         $result = $service->deleteLocation($locationId);
 
         return response()->json(['result' => $result]);
