+++ app/Http/Controllers/ResourceController.php
@@ -1,12 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Controllers;
 
 use App\Http\Requests\IndexResourceRequest;
 use App\Http\Requests\StoreResourceRequest;
 use App\Http\Requests\UpdateResourceRequest;
+use App\Http\Resources\ResourceCollection;
 use App\Http\Resources\ResourceResource;
+use App\Models\Resource;
 use App\Services\ResourceService;
+use Illuminate\Http\JsonResponse;
 
 class ResourceController extends Controller
 {
@@ -15,13 +20,16 @@ class ResourceController extends Controller
      *
      * @param \App\Http\Requests\IndexResourceRequest $request
      * @param \App\Services\ResourceService $service
-     * @return \App\Http\Resources\LocationCollection
+     * @return \App\Http\Resources\ResourceCollection
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function index(IndexResourceRequest $request, ResourceService $service)
+    public function index(IndexResourceRequest $request, ResourceService $service): ResourceCollection
     {
-        $resources = $service->getResourcesByUser($request->user_id);
+        $this->authorize('viewAny', Resource::class);
 
-        return ResourceResource::collection($resources);
+        $resources = $service->getResourcesByUser(auth()->id());
+
+        return new ResourceCollection($resources);
     }
 
     /**
@@ -30,9 +38,12 @@ class ResourceController extends Controller
      * @param \App\Http\Requests\StoreResourceRequest $request
      * @param \App\Services\ResourceService $service
      * @return \App\Http\Resources\ResourceResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function store(StoreResourceRequest $request, ResourceService $service)
+    public function store(StoreResourceRequest $request, ResourceService $service): ResourceResource
     {
+        $this->authorize('create', Resource::class);
+
         $resource = $service->createResource($request->validated());
 
         return new ResourceResource($resource);
@@ -44,9 +55,12 @@ class ResourceController extends Controller
      * @param string $resourceId
      * @param \App\Services\ResourceService $service
      * @return \App\Http\Resources\ResourceResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function show(string $resourceId, ResourceService $service)
+    public function show(string $resourceId, ResourceService $service): ResourceResource
     {
+        $this->authorize('view', [Resource::class, $resourceId]);
+
         $resource = $service->getResourceById($resourceId);
 
         return new ResourceResource($resource);
@@ -59,9 +73,12 @@ class ResourceController extends Controller
      * @param string $resourceId
      * @param \App\Services\ResourceService $service
      * @return \App\Http\Resources\ResourceResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function update(UpdateResourceRequest $request, string $resourceId, ResourceService $service)
+    public function update(UpdateResourceRequest $request, string $resourceId, ResourceService $service): ResourceResource
     {
+        $this->authorize('update', [Resource::class, $resourceId]);
+
         $resource = $service->updateResource($request->validated(), $resourceId);
 
         return new ResourceResource($resource);
@@ -73,9 +90,12 @@ class ResourceController extends Controller
      * @param string $resourceId
      * @param \App\Services\ResourceService $service
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function destroy(string $resourceId, ResourceService $service)
+    public function destroy(string $resourceId, ResourceService $service): JsonResponse
     {
+        $this->authorize('delete', [Resource::class, $resourceId]);
+
         $result = $service->deleteResource($resourceId);
 
         return response()->json(['result' => $result]);
