+++ app/Http/Controllers/ResourceTypeController.php
@@ -1,12 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Controllers;
 
 use App\Http\Requests\IndexResourceTypeRequest;
 use App\Http\Requests\StoreResourceTypeRequest;
 use App\Http\Requests\UpdateResourceTypeRequest;
+use App\Http\Resources\ResourceTypeCollection;
 use App\Http\Resources\ResourceTypeResource;
+use App\Models\ResourceType;
 use App\Services\ResourceTypeService;
+use Illuminate\Http\JsonResponse;
 
 class ResourceTypeController extends Controller
 {
@@ -16,12 +21,15 @@ class ResourceTypeController extends Controller
      * @param \App\Http\Requests\IndexResourceTypeRequest $request
      * @param \App\Services\ResourceTypeService $service
      * @return \App\Http\Resources\ResourceTypeCollection
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function index(IndexResourceTypeRequest $request, ResourceTypeService $service)
+    public function index(IndexResourceTypeRequest $request, ResourceTypeService $service): ResourceTypeCollection
     {
-        $resourceTypes = $service->getResourceTypesByUser($request->user_id);
+        $this->authorize('viewAny', ResourceType::class);
 
-        return ResourceTypeResource::collection($resourceTypes);
+        $resourceTypes = $service->getResourceTypesByUser(auth()->id());
+
+        return new ResourceTypeCollection($resourceTypes);
     }
 
     /**
@@ -30,9 +38,12 @@ class ResourceTypeController extends Controller
      * @param \App\Http\Requests\StoreResourceTypeRequest $request
      * @param \App\Services\ResourceTypeService $service
      * @return \App\Http\Resources\ResourceTypeResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function store(StoreResourceTypeRequest $request, ResourceTypeService $service)
+    public function store(StoreResourceTypeRequest $request, ResourceTypeService $service): ResourceTypeResource
     {
+        $this->authorize('create', ResourceType::class);
+
         $resourceType = $service->createResourceType($request->validated());
 
         return new ResourceTypeResource($resourceType);
@@ -44,9 +55,12 @@ class ResourceTypeController extends Controller
      * @param string $resourceTypeId
      * @param \App\Services\ResourceTypeService $service
      * @return \App\Http\Resources\ResourceTypeResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function show(string $resourceTypeId, ResourceTypeService $service)
+    public function show(string $resourceTypeId, ResourceTypeService $service): ResourceTypeResource
     {
+        $this->authorize('view', [ResourceType::class, $resourceTypeId]);
+
         $resourceType = $service->getResourceTypeById($resourceTypeId);
 
         return new ResourceTypeResource($resourceType);
@@ -59,9 +73,12 @@ class ResourceTypeController extends Controller
      * @param string $resourceTypeId
      * @param \App\Services\ResourceTypeService $service
      * @return \App\Http\Resources\ResourceTypeResource
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function update(UpdateResourceTypeRequest $request, string $resourceTypeId, ResourceTypeService $service)
+    public function update(UpdateResourceTypeRequest $request, string $resourceTypeId, ResourceTypeService $service): ResourceTypeResource
     {
+        $this->authorize('update', [ResourceType::class, $resourceTypeId]);
+
         $resourceType = $service->updateResourceType($request->validated(), $resourceTypeId);
 
         return new ResourceTypeResource($resourceType);
@@ -73,9 +90,12 @@ class ResourceTypeController extends Controller
      * @param string $resourceTypeId
      * @param \App\Services\ResourceTypeService $service
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function destroy(string $resourceTypeId, ResourceTypeService $service)
+    public function destroy(string $resourceTypeId, ResourceTypeService $service): JsonResponse
     {
+        $this->authorize('delete', [ResourceType::class, $resourceTypeId]);
+
         $result = $service->deleteResourceType($resourceTypeId);
 
         return response()->json(['result' => $result]);
